kernelci.lab.shell: run as detached process
Run the script in a detached process and return the PID to follow the
same convention as with LAVA, Kubernetes and other types of lab
running jobs asynchronously.  Alternatively, the extra argument
get_process=True can be used to get the subprocess object rather than
the PID.  This may then be used to check whether the process has
completed etc.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>